pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "realtime-spatial-assistant-slam"
version = "0.1.0"
description = "ORB-SLAM3 Python bindings packaged for realtime_spatial_assistant."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "opencv-python-headless>=4.8",
  "open-clip-torch>=2.24",
  "paddleocr>=2.7",
  "pillow>=10.0",
  "psycopg[binary]>=3.1",
  "qdrant-client>=1.9",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "ultralytics>=8.0",
  "psycopg2>=2.9.11",
]

[tool.setuptools.packages.find]
include = ["database*", "pipeline*", "schema*", "services*", "slam*"]

[tool.uv.workspace]
members = [
    "test_scripts/mcp/mcp-server-demo",
]
